requirements.txt
Django==1.9.4
pycrypto==2.6.1
wheel==0.26.0
